chore: renamed SharpGUI to SharpIDE
