+++ SharpIDE/Assets/avalonia-logo.ico
