+++ SharpIDE/SharpIDE.csproj
