+++ SharpIDE/ViewLocator.cs
@@ -4,7 +4,7 @@ using Avalonia.Controls;
 using Avalonia.Controls.Templates;
 using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace SharpGUI;
+namespace SharpIDE;
 
 /// <summary>
 /// Given a view model, returns the corresponding view if possible.
