+++ SharpIDE/Views/Editor/ScriptEditorView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:AvaloniaEdit="clr-namespace:AvaloniaEdit;assembly=AvaloniaEdit"
-             xmlns:local="clr-namespace:SharpGUI.Views.Editor"
+             xmlns:local="clr-namespace:SharpIDE.Views.Editor"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="SharpGUI.Views.Editor.ScriptEditorView"
+             x:Class="SharpIDE.Views.Editor.ScriptEditorView"
              x:DataType="local:ScriptEditorVM">
   <AvaloniaEdit:TextEditor Name="Editor"
                            Background="#1E1E1E"
