+++ SharpIDE/Views/Main/MainWindow.axaml
@@ -3,14 +3,14 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:dock="clr-namespace:Dock.Avalonia.Controls;assembly=Dock.Avalonia"
-        xmlns:vm="clr-namespace:SharpGUI.Views.Main"
-        xmlns:panels="clr-namespace:SharpGUI.Panels"
-        xmlns:editor="clr-namespace:SharpGUI.Views.Editor"
+        xmlns:vm="clr-namespace:SharpIDE.Views.Main"
+        xmlns:panels="clr-namespace:SharpIDE.Panels"
+        xmlns:editor="clr-namespace:SharpIDE.Views.Editor"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="SharpGUI.Views.Main.MainWindow"
+        x:Class="SharpIDE.Views.Main.MainWindow"
         x:DataType="vm:MainWindowVM"
         Icon="/Assets/avalonia-logo.ico"
-        Title="SharpGUI">
+        Title="SharpIDE">
 
     <Window.DataTemplates>
         <DataTemplate DataType="{x:Type panels:VariableExplorerVM}">
